Добавляет модульность JS-файлов
